unsigned int -> size_t, возможно теперь будет компилиться на 64-битных архитектурах

+++ c++/sdl/sdlapplication/Scene.h
@@ -70,7 +70,7 @@ public:
 		Light.clear();
 	}
 
-	inline unsigned int LightCount() const
+	inline size_t LightCount() const
 	{
 		return Light.size();
 	}
